requirements.txt
# Base numérica
numpy==1.23.5

# OpenCV (evita build com GUI desnecessária)
opencv-python-headless==4.8.1.78

# OpenSlide Python binding
openslide-python==1.3.1

# Pillow
Pillow==10.2.0

# Matplotlib
matplotlib==3.7.5

# TensorFlow (CPU)
tensorflow==2.11.0

# TFLite runtime (mais leve que TF full, opcional mas recomendado)
tflite-runtime==2.11.0
